samfs: number of blocks is stored in inode as int32
on large files, greater than 8TB this value can be negative. In upper OS
filesystem layer this value is assigned to uin32t and turn into a value
larger than MAXLONGINT.
To prevent this, this  change will first cast the int32 to uint32 before it is
assinged to the uint64.

+++ src/fs/samfs/cvnops.c
@@ -1239,7 +1239,7 @@ sam_getattr_vn(
 	}
 	vap->va_blksize = LG_BLK(ip->mp, DD);	/* Block size in bytes */
 	/* # of DEV_BSIZE (512-byte) blocks allocated */
-	vap->va_nblocks = (u_longlong_t)ip->di.blocks *
+	vap->va_nblocks = (u_longlong_t)((uint32_t)ip->di.blocks) *
 	    (u_longlong_t)(SAM_BLK/DEV_BSIZE);
 
 #if defined(AT_SEQ)
